refac(back): #1354 deprecate sandbox
- Deprecate sandbox option from CLI
as it only works on privileged containers
or machines with multi-user nix with
a trusted user executing makes.
Ideally deciding whether or not to use sandbox
should be done by the user via nix.conf.
- Deprecate MAKES_K8S_COMPAT as
it is only used for disabling sandbox,
which is now done by the user via nix.conf.
- Update design principles

Signed-off-by: Daniel Salazar <podany270895@gmail.com>